pyproject.toml
[tool.poetry]
name = "myproject"
version = "0.1.0"
description = ""
authors = ["Van Hung LE <vanhung1512@gmail.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
django = "^3.2"
djangorestframework = "^3.12.4"
psycopg2-binary = "^2.9.1"
pyjwt = "^2.3.0"
bcrypt = "^3.2.0"
django-cors-headers = "^3.8.0"


[tool.poetry.group.dev.dependencies]
setuptools = "^69.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
